Refactor structure of utils files
