+++ evol/helpers/utils.py
@@ -1,63 +1,18 @@
-from inspect import signature
-from typing import Any, Callable, Generator, List, Sequence, Union
-
-from evol import Individual
+from random import randint
+from typing import Tuple
 
 
-def offspring_generator(parents: List[Individual],
-                        parent_picker: Callable[..., Union[Individual, Sequence]],
-                        combiner: Callable[..., Any],
-                        **kwargs) -> Generator[Individual, None, None]:
-    """Generator for offspring.
+def select_partition(length: int, min_size: int = 1, max_size: int = None) -> Tuple[int, int]:
+    """Select a partition of a chromosome.
 
-    This helps create the right number of offspring,
-    especially in the case of of multiple offspring.
-
-    :param parents: List of parents.
-    :param parent_picker: Function that selects parents. Must accept a sequence of
-        individuals and must return a single individual or a sequence of individuals.
-        Must accept all kwargs passed (i.e. must be decorated by select_arguments).
-    :param combiner: Function that combines chromosomes. Must accept a tuple of
-        chromosomes and either return a single chromosome or yield multiple chromosomes.
-        Must accept all kwargs passed (i.e. must be decorated by select_arguments).
-    :param kwargs: Arguments
-    :returns: Children
+    :param length: Length of the chromosome.
+    :param min_size: Minimum length of the partition. Defaults to 1.
+    :param max_size: Maximum length of the partition. Defaults to length - 1.
+    :return: Start and end index of the partition.
     """
-    while True:
-        # Obtain parent chromosomes
-        selected_parents = parent_picker(parents, **kwargs)
-        if isinstance(selected_parents, Individual):
-            chromosomes = (selected_parents.chromosome,)
-        else:
-            chromosomes = tuple(individual.chromosome for individual in selected_parents)
-        # Create children
-        combined = combiner(*chromosomes, **kwargs)
-        if isinstance(combined, Generator):
-            for child in combined:
-                yield Individual(chromosome=child)
-        else:
-            yield Individual(chromosome=combined)
-
-
-def select_arguments(func: Callable) -> Callable:
-    """Decorate a function such that it accepts any keyworded arguments.
-
-    The resulting function accepts any arguments, but only arguments that
-    the original function accepts are passed. This allows keyworded
-    arguments to be passed to multiple (decorated) functions, even if they
-    do not (all) accept these arguments.
-
-    :param func: Function to decorate.
-    :return: Callable
-    """
-
-    def result(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except TypeError:
-            return func(*args, **{k: v for k, v in kwargs.items() if k in signature(func).parameters})
-
-    return result
+    partition_size = randint(min_size, length - 1 if max_size is None else max_size)
+    partition_start = randint(0, length - partition_size)
+    return partition_start, partition_start + partition_size
 
 
 def rotating_window(arr):
